fix: renombramiento y se modifico settings.py

+++ backend/productos/models.py
@@ -4,5 +4,5 @@ class Product(models.Model):
     name = models.CharField(max_length=100)
     quantity = models.FloatField()
     barcode = models.CharField(max_length=50, unique=True, blank=True, null=True)
-    tax = models.DecimalField()
-    price = models.DecimalField()
+    tax = models.FloatField()
+    price = models.FloatField()
